app/requirements.txt
mysql.connector
requests